+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
  - add required `task_type` attribute to OpenquakeHazardSolution (breaking change).
 
+### Changed
+ - make `task_type` attribute on OpenquakeHazardTask `required` (breaking change).
+
 ## [0.4.0] - 2025-12-02
 
 ### Added